Reset marked Exercise on screen rotations.
Ideally we would want to restore the currently marked Exercise when the
screen is rotated, but it will need a bit of investigation on how to
achieve this best. So for now we simply reset the currently selected
Exercise and make sure that the state is at least clean again. Because
then whilst the selection is gone after rotation, at least it can be
selected and interacted with as expected again.