Show app Activities above the lock screen.
This way the user can add Exercises or Sessions without having to
explicitly unlock the screen every time.

Ideally we would show a permanent notification as this is the most
Androidy way of doing it. But, we can implement this in future commits.
For now this is a good compromise.