Update gradle config files after gradle update.
I am actually not entirely sure whether these should be under version
control at all. It certainly feels wrong when configuration files such
as this that depend on the local build environment are committed.
But, as there is only one developer on this project it is not a big
issue for now, and we can have a look into it in the future if we have
to.