Revert "Show app Activities above the lock screen."
This reverts commit 2049ff3eaab8ce538def4627e7a7ec5dbca9f3c0.
Showing the app above the lock screen seems like a great idea in
practice, as it avoids having to unlock the phone every time a session
is added. But, it is a bit glitchy on my Cyanogenmod S3. And, it is
actually a bit annoying as it always shows the app even if you do not
want to see it and just want to check the time or latest messages or so.

So for now it seems like not having this functionality is better than
having it with drawbacks. We will have to consider some better way of
achieving this. Maybe we could create a custom notification to add
sessions for example.