Remove distance attribute from time-based Exercises.
Instead of trying to make everything perfect and having a more
complicated time-based Exercise type, we remove the distance attribute
and simply only measure the time for time-based exercises. This way we
can hopefully make it a feature sooner rather than later.
It is a valid use-case in itself as it can be used for example for
simple core exercises like planks.

In future we can add further types for example specific for running that
include attributes like distance and pace, but for now we start small
and simple and get something simple working first.