Add support for persisting time-based exercise sessions.
So far we were only able to add time-based Exercises, and the individual
Sessions were not stored. With this commit we are adding some basic
support to store the time-based Sessions in our json file. We simply
store and restore the duration of the Session.

Note that we refer to the duration of a Session both as 'duration' and
simply as 'time' depending on where we refer to it. We may consider to
consolidate this at some point in the future to avoid confusion.