Add placeholder for number of session to weight-based exercise sessions.
To make it clear how many sessions were done each day, or how many are
left to go, we will show a counter next to the date.
This commit provides the initial implementation where we add the
TextViews and make space for it. For now it is always hardcoded to 1,
and we will provide an actual implementation for counting in future
commits. This way we will keep the commits to a manageable size.