Set Session numbers when new Sessions are added.
We now set the number of a new Session when it is added. This should
work fine for existing Sessions and when adding new ones.
I haven't quite decided yet though whether showing the numbers makes the
Session list too busy and whether it is distracting. But, we can play
with it for a bit and then decide from there.