Implement mechanism to update range of Sessions after adding new one.
When we insert a new Session in the middle of day, we did not yet have a
way of updating all Sessions after it, and notifying the adapter to
reload it. To do this, we now simply introduce a class that contains
both the position on which the Session was inserted and a range of
Sessions that require updating by the view adapter.

Note that this allows us to get rid of nextSessionNeedsUpdate as well,
as this now falls under the range of Sessions to be updated.