Split ViewExerciseActivity into specific Exercise classes.
So far we have had all logic for the two different exercise types (time
and weight based) implemented in the ViewExerciseActivity class. This is
fine for that case, but we are going to add further Exercise types and
it would make the class get bigger and bigger and more messy. Splitting
it into an abstract base class and implementing more specific Activities
to instantialise seems like a good approach.

Because now all we have to implement in the inheriting classes is two
methods, and the rest is covered by the common main class. There is
still some code duplication between the inheriting classes, but we can
consider refactoring that in future commits. For now this seems like a
good first step towards managing the code a bit more cleanly.