Add explicit storeExercises function.
Eventually we will want to be able to both save exercises to be
automatically backed up and to be exported. Having an explicit function
to just save exercises for now allows us to extend it with an export
function in future commits.

Note that functionality wise this commit does not change anything.