Add basic export action.
We now have an overflow action item to export the current exercises to
external storage. This is pretty basic but should be good enough for
something that is pretty much almost never used.

Note that so far we always store to external storage as we have not
added the internal storage saving functionality yet.