Add stub import exercise action item with confirmation dialog.
As a first step towards proper importing we now have an action item to
do so which prompts a confirmation dialog as well as it is a destructive
action. It does not actually do anything yet, but we will follow up in
future commits to keep the commits to a sane size.