Switch project to more permissive MIT license.
My original intention was to do a bit more with this app, hence the
quite restricted license. I have since changed my mind though and there
is no reason to be as restrictive anymore. With the MIT license
basically everyone can do whatever with the code as long as they mention
the original project.